+++ CHANGELOG.md
@@ -1,10 +1,50 @@
 # voyage-ai-provider
 
-## 1.1.1
+## 2.0.0
+
+### Major Changes
+
+- b7f0922: - update to @ai-sdk/provider and @ai-sdk/provider-utils beta
+  - change to zod v4 internally
 
 ### Patch Changes
 
-- update dependencies
+- 9bf57b7: fix beta versioning
+- b7f0922: - fix the export
+- b7f0922: - export `voyageEmbeddingOptions` and `voyageMultimodalEmbeddingOptions`
+
+## 2.0.0-beta.4
+
+### Patch Changes
+
+- fix beta versioning
+
+## 2.0.0-beta.3
+
+### Patch Changes
+
+- fix the export
+
+## 2.0.0-beta.2
+
+### Patch Changes
+
+- export `voyageEmbeddingOptions` and `voyageMultimodalEmbeddingOptions`
+
+## 2.0.0-beta.1
+
+### Patch Changes
+
+- update to @ai-sdk/provider and @ai-sdk/provider-utils beta
+- change to zod v4 internally
+
+## 2.0.0-alpha.0
+
+### Major Changes
+
+- 2.0.0-alpha.0
+- Upgrade to @ai-sdk/provider v2
+- upgrade to ai V5
 
 ## 1.1.0